requirements.txt
django
django_restframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
psycopg2-binary
dj-database-url